refactor: rewrite function to be more reusable
Change loginUser to createForm that will take a string as argument
and create the form for the login popup. This removes the need
of creating another function for the sign up form.
Also remove overlayDiv as a global variable due to future
purpose of removing createForm and adding it into another
module.