refactor/feat: move function into popupForm.js and add changes
Move function into popupForm module to avoid dependency cycles.
Also add some code that will check if user exists and add them
to localStorage and render logged in options.
And add a condition that will check if there isnt a message
on the popup AND if the user dont exist, then
display the message on the popup.