feat: add function that will validate order form
The first function, validateInputField, is a function
that validates an input. It removes the need of
duplicated code.
The second function, formIsValid, takes a string
as argument. It contains several regex to test
with the input values.
There are three main if statements based on the argument
value, login, signup and order. They each validate
different forms and return several booleans to check
for validity. All booleans need to be true in order for it to return
true.