feat: add select element for category
Separate createHomepageElements into
another function called updateProductDisplay.
This is because we need to update the products
container whenever a new select option
is selected.
Also fetch the categories from API and for each
category, create the option and append
it to the selectElement. The selectElement
has a change eventlistener to empty the productsContainer,
retrieve the filteredProducts and updateProductDisplay.